move src files to src directory
